# walking-safely/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Walking Safely API",
        "description": "API para o aplicativo Walking Safely - Navegação segura com análise de risco em tempo real",
        "contact": {
            "name": "Walking Safely Team",
            "email": "suporte@walkingsafely.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.walkingsafely.com/api",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/routes": {
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Calcular rota",
                "description": "Calcula uma rota entre origem e destino com análise de risco. Pode opcionalmente iniciar uma sessão de navegação.",
                "operationId": "calculateRoute",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin",
                                    "destination"
                                ],
                                "properties": {
                                    "origin": {
                                        "required": [
                                            "latitude",
                                            "longitude"
                                        ],
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -23.5505
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -46.6333
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "destination": {
                                        "required": [
                                            "latitude",
                                            "longitude"
                                        ],
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -23.5629
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -46.6544
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "prefer_safe_route": {
                                        "description": "Preferir rota mais segura",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "start_navigation": {
                                        "description": "Iniciar sessão de navegação",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rota calculada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RouteWithRisk"
                                        },
                                        "occurrences": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Occurrence"
                                            }
                                        },
                                        "session_id": {
                                            "type": "integer",
                                            "example": 123,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Serviço de mapas indisponível",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/routes/recalculate": {
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Recalcular rota",
                "description": "Recalcula a rota durante navegação ativa. Acionado quando tempo de viagem aumenta >10% ou usuário sai da rota.",
                "operationId": "recalculateRoute",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "current_position"
                                ],
                                "properties": {
                                    "session_id": {
                                        "description": "ID da sessão de navegação ativa",
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "current_position": {
                                        "required": [
                                            "latitude",
                                            "longitude"
                                        ],
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -23.555
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -46.64
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rota recalculada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RouteRecalculation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou sessão não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Occurrence": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -23.5505
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -46.6333
                    },
                    "crime_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "crime_type": {
                        "type": "string",
                        "example": "Assalto"
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "critical"
                        ],
                        "example": "medium"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "expired",
                            "rejected",
                            "merged"
                        ],
                        "example": "active"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "collaborative",
                            "official"
                        ],
                        "example": "collaborative"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "confidence_score": {
                        "type": "integer",
                        "example": 3
                    },
                    "region_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "RouteWithRisk": {
                "properties": {
                    "route": {
                        "properties": {
                            "distance": {
                                "description": "Distância em metros",
                                "type": "number",
                                "example": 2500
                            },
                            "duration": {
                                "description": "Duração em segundos",
                                "type": "number",
                                "example": 1800
                            },
                            "polyline": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Coordinates"
                                }
                            },
                            "instructions": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "risk_analysis": {
                        "properties": {
                            "overall_risk": {
                                "type": "string",
                                "enum": [
                                    "low",
                                    "medium",
                                    "high",
                                    "critical"
                                ],
                                "example": "medium"
                            },
                            "risk_score": {
                                "type": "number",
                                "format": "float",
                                "example": 0.45
                            },
                            "risk_segments": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "start_index": {
                                            "type": "integer"
                                        },
                                        "end_index": {
                                            "type": "integer"
                                        },
                                        "risk_level": {
                                            "type": "string"
                                        },
                                        "risk_score": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "risk_factors": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "weight": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "is_safe_route": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "RouteRecalculation": {
                "properties": {
                    "recalculated": {
                        "type": "boolean",
                        "example": true
                    },
                    "reason": {
                        "type": "string",
                        "example": "travel_time_increased"
                    },
                    "new_route": {
                        "$ref": "#/components/schemas/RouteWithRisk"
                    },
                    "risk_changed": {
                        "type": "boolean",
                        "example": false
                    },
                    "previous_risk": {
                        "type": "string",
                        "example": "medium"
                    },
                    "new_risk": {
                        "type": "string",
                        "example": "medium"
                    }
                },
                "type": "object"
            },
...
            "Coordinates": {
                "required": [
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "maximum": 90,
                        "minimum": -90,
                        "example": -23.5505
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "maximum": 180,
                        "minimum": -180,
                        "example": -46.6333
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "validation_error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}